# updated query to use l2_keystones rather than pop_basis in JOIN to improve performance